# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "mygram is a free photo sharing app written in Go. People can share, view, and comment photos by everyone. Anyone can create an account by registering an email address and creating a username.",
        "title": "mygram backend",
        "contact": {
            "name": "gusrylmubarok",
            "email": "gusrylmubarok@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/users": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update a user with authentication user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update a user",
                "parameters": [
                    {
                        "description": "Update User",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseDataUpdatedUser"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseMessage"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseMessage"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete a user with authentication user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Delete a user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessageDeletedUser"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseMessage"
                        }
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "description": "Authentication a user and retrieve a token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Login a user",
                "parameters": [
                    {
                        "description": "Login User",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LoginUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseDataLoggedinUser"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseMessage"
                        }
                    }
                }
            }
        },
        "/users/register": {
            "post": {
                "description": "create and store a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Register a user",
                "parameters": [
                    {
                        "description": "Register User",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RegisterUser"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseDataRegisteredUser"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseMessage"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseMessage"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "helpers.ResponseMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "model.LoggedinUser": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "the token generated here"
                }
            }
        },
        "model.LoginUser": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "johndoe@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "secret"
                }
            }
        },
...
        "model.RegisterUser": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer",
                    "example": 8
                },
                "email": {
                    "type": "string",
                    "example": "johndoe@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "secret"
                },
                "username": {
                    "type": "string",
                    "example": "johndoe"
                }
            }
        },
        "model.RegisteredUser": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer",
                    "example": 8
                },
                "email": {
                    "type": "string",
                    "example": "johndoe@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "the user id generated here"
                },
                "username": {
                    "type": "string",
                    "example": "johndoe"
                }
            }
        },
...
        "model.ResponseDataLoggedinUser": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.LoggedinUser"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "model.ResponseDataRegisteredUser": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.RegisteredUser"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "model.ResponseDataUpdatedUser": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.UpdatedUser"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "model.ResponseMessageDeletedUser": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "your account has been successfully deleted"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "model.UpdateUser": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "newjohndoe@example.com"
                },
                "username": {
                    "type": "string",
                    "example": "newjohndoe"
                }
            }
        },
...
        "model.UpdatedUser": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer",
                    "example": 8
                },
                "email": {
                    "type": "string",
                    "example": "newjohndoe@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "here is the generated user id"
                },
                "updated_at": {
                    "type": "string",
                    "example": "the updated at generated here"
                },
                "username": {
                    "type": "string",
                    "example": "newjohndoe"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Description for what is this security definition being used",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}